Rework sun4i-i2s with codekipper patches for H3s
sun4i-i2s.c and its device tree binding documentation is modified in
order to bring better support for H3 processors and especially NanoPi
NEO in mind. All the changes are made with an inspiration from another
developer's, Marcus Cooper <codekipper@gmail.com>, efforts. He shared
his efforts on linux-sunxi mailing list as a patch of 6 pieces. On this
commit all those commits are put together in order to generate one
single patch file with all the changes and test them on a NanoPi NEO
board.

Therefore, I give all the credit to the original author of the commits,
Marcus Cooper <codekipper@gmail.com> and I do not claim any right on the
effort. The original commits/patches and the corresponding email loop on
linux-sunxi email group can be found on the link below.
https://groups.google.com/forum/#!topic/linux-sunxi/rTn02ry5Qmo