Fix offset tune for SUN8I \n SUN8I_I2S_TX_CHAN_OFFSET_MASK should be GENMASK(13,12) And offset tune of rx should also be updated.
Fixes: 7d2993811a1e ("ASoC: sun4i-i2s: Add support for H3")

This commit is just applies a patch shared on linux-sunxi mailinglist.
Therefore, I do not claim any ownership on this effort. Sole purpose of
this commit is to create patch files to test untested/unmerged commits
in order to make H3 i2s working on NanoPi NEO board.

The original commit and patch file can be found on the link below:
https://groups.google.com/forum/#!topic/linux-sunxi/7SCM4n8LChE